chore: Remove app bar title in home page and add references to invoice form
This commit removes the app bar title in the home page of the EASEHomePage widget. It also adds the \References\ section to the invoice form in the ManageReferencesWidget, allowing users to input reference information for their invoices. These changes improve the user experience and provide more functionality in the app.