refactor: Update Invoice model to use nullable ID
The Invoice model in invoice.dart is updated to use a nullable ID by changing the type of the 'id' property from 'int' to 'int?'. This change allows for auto-incrementing IDs from the database and makes the ID field a primary key. The purpose of this change is to enhance the functionality and flexibility of the Invoice model.